+++ IntuneWinPackager.Core/Interfaces/IPreflightService.cs
@@ -1,0 +1,8 @@
+using IntuneWinPackager.Models.Entities;
+
+namespace IntuneWinPackager.Core.Interfaces;
+
+public interface IPreflightService
+{
+    Task<PreflightResult> RunAsync(PackagingRequest request, CancellationToken cancellationToken = default);
+}