+++ IntuneWinPackager.Models/Enums/PreflightSeverity.cs
@@ -1,0 +1,8 @@
+namespace IntuneWinPackager.Models.Enums;
+
+public enum PreflightSeverity
+{
+    Info = 0,
+    Warning = 1,
+    Error = 2
+}